Remove kludge
Phaser has a lot of ways to do the same thing. Previously I found one that I thought was right (bounds via camera)
but that led to an initialization problem. On rethinking (and testing what works and doesn't in phaser), it makes
more sense to test boundaries with the physics world. While the camera itself should never change views and scroll
with the player (or anything else), the world itself is a more semantically correct check for player and obstacle
death.

Note to future self. I may never read this again, but writing things like this is helpful for my memory.